[ios][precompile][spm] make emc work with swift package manager
This is the first of a series of PRs to adjust ExpoModulesCore to build using SPM - ie. through a definition in a Package.swift.

Another PR will add support for generating, building and packaging EMC through expotools.

This PR contains the simplest set of changes for SPM to be able to build EMC and is mainly about trying to not re-expose symbols (headers) from React Native's prebuilt binaries (xcframeworks).

The main problem that this PR tries to solve is to remove exports of React Native headers. This is a bit tricky when we have classes that inherits from a React Native type - then we need to export the React Native types by including the header files since forward declaring types use for inheritance is not possible in Objective-c.

- Deleted the RCTComponentDataSwiftAdapter since it is no longer used in the new architecture
- Removed use of EXHostWrapper
- Removed re-export of RCTHTTPRequestHandler in ExpoModulesCore.h
- Resolved imports of umbrella headers so that they use the correct paths depending on which "mode" we're building in
- Used built-in macro __building_module combined with forward declaration to avoid re-exporting React headers
- Removed reference to Swift types that hasn't been built when building obj-c with SPM (EXFileSystemLegacyUtilities in EXModuleRegistryProvider.m)
- Removed direct import of Swift interface from obj-c (it isn't built yet when we build the obj-c target)